requirements.txt
# Flask Backend
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
bcrypt==4.1.1

# Telegram Bot (with webhook support)
python-telegram-bot[webhooks]==21.0

# HTTP requests
requests==2.31.0

# Excel export
pandas>=2.2.3
openpyxl==3.1.2

# Printer support
python-escpos==3.0
reportlab==4.0.7
Pillow==10.1.0

# WSGI server for production
gunicorn==21.2.0
